feat: finish part 1(html, validation, file upload)
